fix VM tests
