+++ packages/ui/ViewManager/tests/View-specs.js
@@ -3,34 +3,9 @@ import React from 'react';
 import {mount} from 'enzyme';
 import View from '../View';
 
-describe('ViewManager', () => {
+import {MockArranger} from './test-utils';
 
-	function MockAnimation (duration) {
-		this.playbackState = 'running';
-		this.finish = () => {
-			if (this.playState === 'finished') return;
-
-			this.playState = 'finished';
-			if (this.onfinish) {
-				this.onfinish();
-			}
-		};
-
-		this.cancel = () => {
-			this.playState = 'finished';
-			if (this.oncancel) {
-				this.oncancel();
-			}
-		};
-
-		setTimeout(this.finish, duration);
-	}
-
-	const MockArranger = {
-		enter: ({duration}) => new MockAnimation(duration),
-		stay: ({duration}) => new MockAnimation(duration),
-		leave: ({duration}) => new MockAnimation(duration)
-	};
+describe('View', () => {
 
 	test(
 		'should render its child when neither enteringProp or childProps is specified',
